Renamed photini.exiv2 module
