+++ src/photini/filemetadata.py
