+++ src/photini/ownership.py
@@ -21,7 +21,7 @@ from __future__ import unicode_literals
 from datetime import datetime
 import logging
 
-from photini.exiv2 import ImageMetadata
+from photini.filemetadata import ImageMetadata
 from photini.pyqt import (
     catch_all, MultiLineEdit, QtCore, QtSlot, QtWidgets, SingleLineEdit)
